requirements.txt
aniso8601==9.0.1
anyio==3.7.0
astunparse==1.6.3
bcrypt==4.0.1
certifi==2023.5.7
cfgv==3.3.1
charset-normalizer==3.1.0
click==8.1.3
distlib==0.3.6
dnspython==2.3.0
exceptiongroup==1.1.2
fastapi==0.99.1
filelock==3.12.2
graphene==3.2.2
graphql-core==3.2.3
graphql-relay==3.2.0
h11==0.14.0
identify==2.5.24
idna==3.4
iniconfig==2.0.0
nodeenv==1.8.0
packaging==23.1
platformdirs==3.8.0
pluggy==1.2.0
pre-commit==3.3.3
pydantic==1.10.11
pymongo==4.4.0
pytest==7.4.0
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0
requests==2.31.0
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
strawberry-graphql==0.193.0
tomli==2.0.1
typing-extensions==4.7.1
urllib3==2.0.3
uvicorn==0.22.0
virtualenv==20.23.1
